Move resource methods from e2e fw to e2e resource fw
