+++ test/e2e/framework/BUILD
@@ -70,7 +70,6 @@ go_library(
         "//test/e2e/framework/metrics:go_default_library",
         "//test/e2e/framework/node:go_default_library",
         "//test/e2e/framework/pod:go_default_library",
-        "//test/e2e/framework/resource:go_default_library",
         "//test/e2e/framework/ssh:go_default_library",
         "//test/e2e/system:go_default_library",
         "//test/utils:go_default_library",
