+++ test/e2e/framework/autoscaling/BUILD
@@ -17,6 +17,7 @@ go_library(
         "//test/e2e/framework:go_default_library",
         "//test/e2e/framework/kubectl:go_default_library",
         "//test/e2e/framework/rc:go_default_library",
+        "//test/e2e/framework/resource:go_default_library",
         "//test/e2e/framework/service:go_default_library",
         "//test/utils:go_default_library",
         "//test/utils/image:go_default_library",
